, fix shard_count in data node's caches + change types, that can hold values > int::max commit_hash:e1d43cd6242b0a01a4593bf8f87babf82b13191d

+++ yt/yt/server/node/data_node/config.h
@@ -61,8 +61,8 @@ struct TP2PConfig
     TSlruCacheDynamicConfigPtr RequestCacheOverride;
 
     TDuration ChunkCooldownTimeout;
-    int MaxDistributedBytes;
-    int MaxBlockSize;
+    i64 MaxDistributedBytes;
+    i64 MaxBlockSize;
     int BlockCounterResetTicks;
     int HotBlockThreshold;
     int SecondHotBlockThreshold;
